Erosion now works as is intended. Talos is now a largely irrelevant variable now (by the way I found out it was spelt Talus, whoops) and taking its place will be a variable called erosionIterations which is simply how many times the algorithm will be run. There is a small kink that needs to be worked out with erosion - sometimes the edges of the terrain are slightly higher than what would seem normal.

+++ include/algorithms/HeightMapGen.h
@@ -29,7 +29,7 @@ class HeightMapGen
 
         void genQuadrant(int xNW, int yNW, int xSE, int ySE, int iteration, int quadrant);
         void erodeBlock(float c, long unsigned int block);
-        void transportMaterial(float material, long unsigned int block);
+        void transportMaterial(unsigned short material, long unsigned int block);
         float getHeightDifference(unsigned long inspectedBlock, unsigned long neighbourBlock);
         void writeMap();
